+++ ui/app/src/CHANGELOG.md
@@ -4,6 +4,7 @@
 * [components]
   * `SearchBar`: Update the default DecoratorIcon to select the search input on click.
   * `PathNavigator/PathNavigatorTree`: Updated `label` prop to be of type `TranslationOrText`.
+  * `PathNavigatorTreeUI/PathNavigatorUI/PathNavigatorTreeHeader`: Updated `title` prop to be of type `string | ReactNode`.
 
 ## 5.0.0 (CrafterCMS 4.4.3)
 * Version 5.0.0 should be paired with CrafterCMS 4.4.3 or later.